+++ NEWS.md
@@ -1,4 +1,5 @@
 # 1.5.0-9003
+* Replaced DefinePSUSamplingParameters (processdata) with ComputePSUSamplingParameters (no processdata) and ReadPSUSamplingParameters (no processdata). This change breaks some pre-release projects (v. v1.3-9006). (#127)
 * Replaced DefineIndividualSamplingParameters (processdata) with ComputeIndividualSamplingParameters (no processdata). This change breaks some pre-release projects (v. v1.3-9006).
 * Added documentation for ECA convergence analsysis (#153)
 * Fixed naming of age groups in reports from analytical estimates (#145)